Moved files, updated paths, as auto not working
