+++ mbs_results/outlier_detection/winsorisation.py
@@ -1,9 +1,9 @@
 import pandas as pd
 
-from mbs_results.calculate_predicted_unit_value import calculate_predicted_unit_value
-from mbs_results.calculate_ratio_estimation import calculate_ratio_estimation
-from mbs_results.calculate_winsorised_weight import calculate_winsorised_weight
-from mbs_results.flag_for_winsorisation import winsorisation_flag
+from mbs_results.outlier_detection.calculate_predicted_unit_value import calculate_predicted_unit_value
+from mbs_results.outlier_detection.calculate_ratio_estimation import calculate_ratio_estimation
+from mbs_results.outlier_detection.calculate_winsorised_weight import calculate_winsorised_weight
+from mbs_results.outlier_detection.flag_for_winsorisation import winsorisation_flag
 
 
 def winsorise(
