+++ mbs_results/outputs/growth_rate_output.py
