+++ old_new_structure.md
@@ -44,18 +44,18 @@ mbs_results/
 │   ├── calculate_imputation_link.py
 │   ├── cumulative_imputation_links.py
 │   ├── flag_and_count_matched_pairs.py
-│   ├── imputation_flags.py
-│   ├── link_filter.py
+│   ├── imputation_flags.pyk
+│   ├── link_filter.py          #1 function but 2 TODOs 
 │   ├── pivot_imputation_value.py
 │   ├── predictive_variable.py
-│   ├── ratio_of_means.py
+│   ├── ratio_of_means.py       # alot of commented out lines and TODOs
 │   └── validate_imputation.py
 ├── estimation/
 │   ├── `__init__.py`
-│   ├── apply_imputation_link.py
+│   ├── apply_imputation_link.py    #  Alot of inconsistant comments and TODOs
 │   ├── calculate_estimation_weights.py
 │   ├── pre_processing_estimation.py
-│   └── validate_estimation.py
+│   └── validate_estimation.py      #  line 49/50 comments
 ├── construction/
 │   ├── `__init__.py`
 │   └── construction_matches.py
@@ -72,12 +72,12 @@ mbs_results/
 │   └── growth_rate_output.py
 ├── utilities/
 │   ├── utils.py
-│   ├── data_cleaning.py
+│   ├── data_cleaning.py    # Alot of comments and functions, potential for cleaning up
 │   └── inputs.py
 └── unsorted/
-    ├── constrains.py - applied after imputation?
+    ├── constrains.py - applied after imputation?   #Alot of comments including operation changes
     ├── selective_editing.py?
-    ├── turnover_analysis.py
+    ├── turnover_analysis.py    #Comment on lines 70-72 seems important?
     └── validation_checks.py
 
 test/
